Apagando um evento da aplicação
